refactor: altera organização dos componentes das dashboard
