+++ src/pages/Dashboard/Dashboard.view.tsx
@@ -1,16 +1,16 @@
-import Collumns from "./components/Columns";
 import * as S from "./Dashboard.styles";
-import { SearchBar } from "./components/Searchbar";
 import { useDashboardController } from "./Dashboard.controller";
+import { Searchbar, Columns } from "./components";
 
 const DashboardPage = () => {
   const { registrations } = useDashboardController();
 
   return (
     <S.Container>
-      <SearchBar />
-      <Collumns registrations={registrations} />
+      <Searchbar />
+      <Columns registrations={registrations} />
     </S.Container>
   );
 };
+
 export default DashboardPage;
