+++ src/pages/Dashboard/components/Columns/Columns.styles.tsx
@@ -1,10 +1,12 @@
 import styled from "styled-components";
+import { RegistrationStatus } from "~/common/interfaces/Registration";
+
 const registrationStatusStyles: {
-  [key in string]: { background: string; title: string };
+  [key in RegistrationStatus]: { background: string; title: string };
 } = {
   REVIEW: {
     background: "#FDF8E9",
-    title: "#EFC24D",
+    title: "#d4a01d",
   },
   APPROVED: {
     background: "#EEEEFD",
@@ -24,7 +26,7 @@ export const Container = styled.div`
   margin-top: 24px;
 `;
 
-export const Column = styled.div<{ status: any }>`
+export const Column = styled.div<{ status: RegistrationStatus }>`
   height: auto;
   background-color: ${({ status }) =>
     registrationStatusStyles[status].background};
@@ -33,7 +35,7 @@ export const Column = styled.div<{ status: any }>`
   max-height: 80vh;
 `;
 
-export const TitleColumn = styled.h3<{ status: any }>`
+export const TitleColumn = styled.h3<{ status: RegistrationStatus }>`
   margin: 0px;
   color: ${({ status }) => registrationStatusStyles[status].title};
   margin: 24px;
