+++ src/pages/Dashboard/components/Columns/Columns.types.ts
@@ -1,0 +1,13 @@
+import {
+  Registration,
+  RegistrationStatus,
+} from "~/common/interfaces/Registration";
+
+export type Column = {
+  status: RegistrationStatus;
+  title: string;
+};
+
+export interface ColumnsProps {
+  registrations?: Registration[];
+}
