+++ src/pages/Dashboard/components/RegistrationCard/RegistrationCard.view.tsx
@@ -1,0 +1,41 @@
+import * as S from "./RegistrationCard.styles";
+import {
+  HiOutlineMail,
+  HiOutlineUser,
+  HiOutlineCalendar,
+  HiOutlineTrash,
+} from "react-icons/hi";
+import { Registration } from "~/common/interfaces/Registration";
+import { IconButton } from "~/components";
+
+const RegistrationCardView = ({ data }: { data: Registration }) => {
+  return (
+    <S.Card>
+      <S.IconAndText>
+        <HiOutlineUser />
+        <h3>{data.employeeName}</h3>
+      </S.IconAndText>
+      <S.IconAndText>
+        <HiOutlineMail />
+        <p>{data.email}</p>
+      </S.IconAndText>
+      <S.IconAndText>
+        <HiOutlineCalendar />
+        <span>{data.admissionDate}</span>
+      </S.IconAndText>
+      <S.Actions>
+        <div className="managing-buttons">
+          <S.ButtonSmall bgcolor="rgb(255, 145, 154)">Reprovar</S.ButtonSmall>
+          <S.ButtonSmall bgcolor="rgb(155, 229, 155)">Aprovar</S.ButtonSmall>
+          <S.ButtonSmall bgcolor="#ff8858">Revisar novamente</S.ButtonSmall>
+        </div>
+
+        <IconButton className="delete-button">
+          <HiOutlineTrash />
+        </IconButton>
+      </S.Actions>
+    </S.Card>
+  );
+};
+
+export default RegistrationCardView;
