+++ src/pages/Dashboard/components/index.ts
@@ -1,0 +1,3 @@
+export { default as RegistrationCard } from "./RegistrationCard/RegistrationCard.view";
+export { default as Searchbar } from "./Searchbar/Searchbar.view";
+export { default as Columns } from "./Columns/Columns.view";
